Added displayed data limit to table 1. Now table has limit for default displayed data and 'show more' button to increase it.
