+++ src/client/app/components/common/editButton.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { penpaperIcon } from '../../../assets/icons';
 
 export default function EditButton({ onClick }) {
-  return <button className='icon-button' onClick={onClick} type='button'>{penpaperIcon}</button>;;
+  return <button aria-roledescription='Редактировать' className='icon-button' onClick={onClick} type='button'>{penpaperIcon}</button>;;
 };
 EditButton.propTypes = {
   onClick: PropTypes.func.isRequired,
