--- src/client/app/components/common/table/tableBody.jsx
@@ -1,43 +1,0 @@
-/* eslint-disable react/forbid-prop-types */
-import React from 'react';
-import lod from 'lodash';
-import PropTypes from 'prop-types';
-// Утилиты
-
-export default function TableBody({ data, columns }) {
-  function renderContent(item, column) {
-    if (columns[column].component) {
-      const { component } = columns[column];
-      if (typeof (component) === 'function') {
-        return component(item);
-      }
-      return component;
-    }
-    return lod.get(item, columns[column].path);
-  }
-  return (
-    <tbody>
-      {data.map(
-        (
-          item, // item - объект одного пользователя
-        ) => (
-          <tr key={item._id}>
-            {Object.keys(columns).map(
-              (
-                column, // columns - объект с информацией для отображения и сортировки
-              ) => (
-                <td key={column}>
-                  {renderContent(item, column)}
-                </td>
-              ),
-            )}
-          </tr>
-        ),
-      )}
-    </tbody>
-  );
-}
-TableBody.propTypes = {
-  columns: PropTypes.object.isRequired,
-  data: PropTypes.array.isRequired,
-};
